Fix infinite loop in prod builds after the `browserslist` update
In commit a4820fc, `browserslist` was updated in order to support `BigInt` with `**` operator. This caused infinite loop in production builds because decorators were no longer being transpiled as functions, the function check was added in 6c45011 to the Terser minification plugin with `keep_fnames` option but `keep_classnames` were intentionally commented out because it had no effect.

This commit enables the `keep_classnames` option to exclude minification of classes that start with "_Pg".